feat: 사다리 게임을 실행하는 기능 구현
