+++ README.md
@@ -35,8 +35,8 @@
 ### 사다리 게임
 
 - [ ] 실행 결과를 입력 받는다.
-- [ ] 사다리 게임을 실행한다
-    - [ ] Line 만큼 반복한다
+- [x] 사다리 게임을 실행한다
+    - [x] Line 만큼 반복한다
         - [x] 플레이어 기준 움직일 수 있는 바의 위치를 확인한다.
             - [x] 플레이어 기준 왼쪽에 있으면 위치를 -1 한다.
             - [x] 플레이어 기준 오른쪽에 있으면 위치를 +1 한다.
