+++ src/test/java/ladder/domain/LadderTest.java
@@ -2,8 +2,12 @@ package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.ArrayList;
+import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 class LadderTest {
 
@@ -15,5 +19,22 @@ class LadderTest {
         Ladder ladder = Ladder.generate(3, 4, new RandomBooleanGenerator());
         //then
         assertThat(ladder.getLines()).hasSize(3);
+    }
+
+    /*
+    |---|   |
+    |   |---|
+    */
+    @ParameterizedTest
+    @CsvSource(value = {"0:2", "1:0", "2:1"}, delimiter = ':')
+    @DisplayName("출발 위치에 따라 결과를 반환한다")
+    void shouldReturnResultPositionWhenMoveLadder(int startPosition, int expectPosition) {
+        //given
+        List<Boolean> determinedBars = new ArrayList<>(List.of(true, false, true));
+        Ladder ladder = Ladder.generate(2, 2, new DeterminedBooleanGenerator(determinedBars));
+        //when
+        Position resultPosition = ladder.getResultPosition(new Position(startPosition));
+        //then
+        assertThat(resultPosition).isEqualTo(new Position(expectPosition));
     }
 }
